optimize ordinary text parsing
Previously, the parser would for every ordinary character

- spawn a list of special characters (allocation is costly)
- perform  with the current character on that list (has O(n) time complexity)

Because this is ran for most characters, this was a bottleneck. Writing out an explicit function that checks
for all the cases is tedious, but much more efficient. In my benchmark, performance for a couple paragraphs went from ~170 runs/s to ~7000 runs/s